@@ -1,5 +1,5 @@
 {
-  "name": "CopaFilmes.SPA",
+  "name": "Friends.SPA",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
@@ -14,17 +14,20 @@
   "private": true,
   "dependencies": {
     "@angular/animations": "^5.2.0",
+    "@angular/cdk": "^5.2.5",
     "@angular/common": "^5.2.0",
     "@angular/compiler": "^5.2.0",
     "@angular/core": "^5.2.0",
     "@angular/forms": "^5.2.0",
     "@angular/http": "^5.2.0",
+    "@angular/material": "^5.2.5",
     "@angular/platform-browser": "^5.2.0",
     "@angular/platform-browser-dynamic": "^5.2.0",
     "@angular/platform-server": "^5.2.0",
     "@angular/router": "^5.2.0",
     "@nguniversal/module-map-ngfactory-loader": "^5.0.0-beta.5",
     "admin-lte": "2.3.11",
+    "ajv": "^6.5.4",
     "angular-font-awesome": "^3.1.2",
     "aspnet-prerendering": "^3.0.1",
     "bootstrap": "^4.1.3",
@@ -33,6 +36,7 @@
     "font-awesome": "^4.7.0",
     "https-proxy-agent": "^2.2.1",
     "jquery": "^3.3.1",
+    "popper.js": "^1.14.4",
     "rxjs": "^5.5.6",
     "sweetalert2": "^7.28.8",
     "zone.js": "^0.8.19"
@@ -52,10 +56,10 @@
     "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.1.2",
+    "protractor": "^5.4.1",
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
-    "typescript": "^2.5.3"
+    "typescript": "^3.1.5"
   },
   "optionalDependencies": {
     "node-sass": "^4.9.0"
