@@ -1,38 +1,50 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
-import { HttpClientModule } from '@angular/common/http';
+import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
 import { RouterModule, PreloadAllModules } from '@angular/router';
 import { HttpModule } from '@angular/http';
+import { ReactiveFormsModule } from '@angular/forms';
+
+import { ErrorInterceptor } from './helpers/error.interceptor';
+import { JwtInterceptor } from './helpers/jwt.interceptor';
+import { AuthGuard } from './auth/auth.guard';
 
 import { ROUTES } from './app.routes';
 
 import { AppComponent } from './app.component';
 import { HomeComponent } from './home/home.component';
-//import { FilmesComponent } from './filmes/filmes.component';
-//import { FilmeComponent } from './filmes/filme/filme.component';
+import { AmigosComponent } from './amigos/amigos.component';
 import { HeaderComponent } from './header/header.component';
+import { LoginComponent } from './login/login.component';
 
-//import { FilmeService } from './filmes/filme/filme.service';
-//import { FilmesService } from './filmes/filmes.service';
+import { AmigosService } from './amigos/amigos.service';
+import { LoginService } from './login/login.service';
 
 
 @NgModule({
   declarations: [
     AppComponent,
     HomeComponent,
-    //FilmesComponent,
-    //FilmeComponent,
+    AmigosComponent,
     HeaderComponent,
+    LoginComponent
   ],
   imports: [
     BrowserModule.withServerTransition({ appId: 'ng-cli-universal' }),
     HttpClientModule,
     HttpModule,
     FormsModule,
+    ReactiveFormsModule,
     RouterModule.forRoot(ROUTES, { preloadingStrategy: PreloadAllModules })
   ],
-  providers: [], //FilmesService, FilmeService, PlacarService],
+  providers: [
+    AmigosService,
+    LoginService,
+    AuthGuard,
+    { provide: HTTP_INTERCEPTORS, useClass: JwtInterceptor, multi: true },
+    { provide: HTTP_INTERCEPTORS, useClass: ErrorInterceptor, multi: true }
+  ],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
