@@ -168,12 +168,83 @@ namespace Friends.Repository
             return retorno;
         }
 
+        private IEnumerable<ContatosViewModel> GetCloseFriends(int id)
+        {
+            JsonSerializer serializer = new JsonSerializer();
+            List<Contatos> listContatos = new List<Contatos>();
+            List<ContatosViewModel> retorno = new List<ContatosViewModel>();
+
+            string sQuery = "select c.*,e.* " +
+                            "from contatos c " +
+                            "inner join enderecos e on c.id_endereco = e.id";
+
+            var result = Connection.Query<Contatos, Enderecos, Contatos>(sQuery,
+                map: (contatos, enderecos) =>
+                {
+                    contatos.endereco = enderecos;
+                    return contatos;
+                },
+                splitOn: "id_endereco,id",
+                transaction: Transaction);
+
+            foreach (var item in result)
+            {
+                if ((string.IsNullOrEmpty(item.endereco.Longitude)) || (string.IsNullOrEmpty(item.endereco.Latitude)))
+                    listContatos.Add(GetLatLongtoAddress(item));
+                else
+                    listContatos.Add(item);
+            }
+
+            Contatos porigem = listContatos.Where(p => p.Id == id).FirstOrDefault();
+
+            var res = from tc in listContatos
+                      select new
+                      {
+                          Contatos = tc,
+                          distance = GetDistanceLatLong(porigem, tc),
+                      };
+
+            var t = res.OrderBy(p => p.distance).Where(p => p.distance > 0).Take(3).ToList();
+
+            //InsertHistoricoLog(porigem.Id)
+
+            foreach (var item in t)
+            {
+                retorno.Add(new ContatosViewModel()
+                {
+                    Cidade = item.Contatos.endereco.Cidade,
+                    Distancia = Math.Round(double.Parse(item.distance.ToString())).ToString(),
+                    Estado = item.Contatos.endereco.Estado,
+                    Logradouro = item.Contatos.endereco.Logradouro,
+                    Nome = item.Contatos.Nome,
+                    Numero = item.Contatos.endereco.Numero,
+                    Pais = item.Contatos.endereco.Pais,
+                    Sobrenome = item.Contatos.Sobrenome
+                });
+
+                InsertHistoricoLog(porigem.Id, item.Contatos.Id, item.distance.ToString());
+            }
+
+            return retorno;
+        }
+
         public double GetDistanceLatLong(Contatos porigem, Contatos pdestino)
         {
             var sCoord = new GeoCoordinate(double.Parse(porigem.endereco.Latitude, CultureInfo.InvariantCulture), double.Parse(porigem.endereco.Longitude, CultureInfo.InvariantCulture));
             var eCoord = new GeoCoordinate(double.Parse(pdestino.endereco.Latitude, CultureInfo.InvariantCulture), double.Parse(pdestino.endereco.Longitude, CultureInfo.InvariantCulture));
 
             return sCoord.GetDistanceTo(eCoord);
+        }
+
+        public async Task<IEnumerable<ContatosListViewModel>> SelectList()
+        {
+            string sQuery = "SELECT c.id, c.nome+ ' ' +c.Sobrenome AS nome " +
+                            "FROM Contatos c ";
+
+            var result = await Connection.QueryAsync<ContatosListViewModel>(
+                sQuery, transaction: Transaction);
+
+            return result.ToList();
         }
 
         public async Task<Contatos> GetByID(int id)
@@ -199,7 +270,7 @@ namespace Friends.Repository
 
         public async Task<IEnumerable<Contatos>> SelectAll()
         {
-            string sQuery = "SELECT c.id, c.nome, c.Sobrenome, e.* " +
+            string sQuery = "SELECT c.id, id_endereco, c.nome, c.sobrenome, e.* " +
                             "FROM Contatos c inner join Enderecos e on c.id_Endereco = e.id";
 
             var result = await Connection.QueryAsync<Contatos, Enderecos, Contatos>(
@@ -218,6 +289,12 @@ namespace Friends.Repository
         public List<ContatosViewModel> GetCloseContacts(string nome, string sobrenome)
         {
             var result = GetCloseFriends(nome, sobrenome);
+            return result.ToList();
+        }
+
+        public List<ContatosViewModel> GetCloseContacts(int id)
+        {
+            var result = GetCloseFriends(id);
             return result.ToList();
         }
     }
