@@ -1,4 +1,4 @@
 export const environment = {
   production: true,
-  api: 'https://copafilmes.azurewebsites.net/api'
+  api: 'http://localhost:7627'
 };
