@@ -1,3 +1,3 @@
-@using CopaFilmes.SPA
-@namespace CopaFilmes.SPA.Pages
+@using Friends.SPA
+@namespace Friends.SPA.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
